feat(cli): unify prepare/prove/verify with zero-arg defaults
- Merge the former `compile` subcommand into `prepare`. Running
  `provekit-cli prepare` now compiles the enclosing Nargo package and
  writes `<circuit>.pkp` / `<circuit>.pkv`. An optional positional
  `[<program_dir>]` selects a different project, and `-p/--pkp` /
  `-v/--pkv` override the default output paths.
- Skip the JSON round-trip in `prepare`: thread the in-memory
  `ProgramArtifact` directly into `NoirProofScheme::from_program`
  instead of re-reading the file we just wrote.
- Convert `prove` and `verify` positionals to flags. Default paths
  (`<circuit>.pkp`, `<circuit>.pkv`, `./Prover.toml`, `./proof.np`) are
  resolved from the enclosing Nargo package, so both commands work with
  zero arguments. Overrides: `-p/-i/-o` for prove, `-v/--proof` for
  verify.
- Extract `enclosing_package_name` and `resolve_key_path` into
  `tooling/cli/src/cmd/util.rs`.
