+++ README.md
@@ -14,30 +14,34 @@ noirup --version v1.0.0-beta.11
 
 > _NOTE:_ The example below is being run for single example `poseidon-rounds`. You can use different example to run same commands.
 
-Compile the Noir circuit:
+Compile the Noir circuit and generate prover/verifier files:
 
 ```sh
 cd noir-examples/poseidon-rounds
-nargo compile
+cargo run --release --bin provekit-cli prepare
 ```
 
-Prepare the Noir program (generates prover and verifier files):
+This compiles the package in the current directory and writes `<circuit>.pkp` and `<circuit>.pkv` next to it. You can also pass an explicit project directory and override the output paths:
 
 ```sh
-cargo run --release --bin provekit-cli prepare ./target/basic.json --pkp ./prover.pkp --pkv ./verifier.pkv
+cargo run --release --bin provekit-cli prepare ./noir-examples/poseidon-rounds --pkp ./prover.pkp --pkv ./verifier.pkv
 ```
 
 Generate the Noir Proof using the input Toml:
 
 ```sh
-cargo run --release --bin provekit-cli prove ./prover.pkp ./Prover.toml -o ./proof.np
+cargo run --release --bin provekit-cli prove
 ```
+
+This reads `<circuit>.pkp` and `./Prover.toml` from the current directory and writes `./proof.np`. Override any of these with `-p`/`--prover`, `-i`/`--input`, or `-o`/`--out`.
 
 Verify the Noir Proof:
 
 ```sh
-cargo run --release --bin provekit-cli verify ./verifier.pkv ./proof.np
+cargo run --release --bin provekit-cli verify
 ```
+
+This reads `<circuit>.pkv` and `./proof.np` from the current directory. Override with `-v`/`--verifier` or `--proof`.
 
 Generate inputs for Gnark circuit:
 
@@ -80,8 +84,8 @@ Benchmark against Barretenberg:
 
 ```sh
 cd noir-examples/poseidon-rounds
-cargo run --release --bin provekit-cli prepare ./target/basic.json --pkp ./prover.pkp --pkv ./verifier.pkv
-hyperfine 'nargo execute && bb prove -b ./target/basic.json -w ./target/basic.gz -o ./target' '../../target/release/provekit-cli prove ./prover.pkp ./Prover.toml'
+cargo run --release --bin provekit-cli prepare --pkp ./prover.pkp --pkv ./verifier.pkv
+hyperfine 'nargo execute && bb prove -b ./target/basic.json -w ./target/basic.gz -o ./target' '../../target/release/provekit-cli prove -p ./prover.pkp -i ./Prover.toml'
 ```
 
 ### Profiling
@@ -92,7 +96,7 @@ The `provekit-cli` application has written custom memory profiler that prints ba
 runs. To run binary with profiling enabled run it with cargo `--features profiling` param or compile with it.
 
 ```sh
-cargo run --release --bin provekit-cli --features profiling prove ./prover.pkp ./Prover.toml -o ./proof.np
+cargo run --release --bin provekit-cli --features profiling prove -p ./prover.pkp -i ./Prover.toml -o ./proof.np
 ```
 
 #### Using tracy (CPU and Memory usage)
@@ -130,7 +134,7 @@ cargo build --release --bin provekit-cli --features profiling
 ```
 5. Now start the application to profile:
 ```sh
-../../target/release/provekit-cli prove ./prover.pkp ./Prover.toml -o ./proof.np
+../../target/release/provekit-cli prove -p ./prover.pkp -i ./Prover.toml -o ./proof.np
 ```
 6. Go back to tracy tool. You should see that it receives data. App is interactive.
 
@@ -141,7 +145,7 @@ open a webpage with interactive app to view results. This does not require to ru
 with profiling enabled.
 
 ```sh
-samply record -r 10000 -- ./../../target/release/provekit-cli prove ./prover.pkp ./Prover.toml -o ./proof.np
+samply record -r 10000 -- ./../../target/release/provekit-cli prove -p ./prover.pkp -i ./Prover.toml -o ./proof.np
 ```
 
 #### Using instruments (Memory usage) - OSX only
@@ -150,7 +154,7 @@ Cargo instruments tool [website](https://crates.io/crates/cargo-instruments) wit
 results using built-in Instruments app. Results are interactive.
 
 ```sh
-cargo instruments --template Allocations --release --bin provekit-cli prove ./prover.pkp ./Prover.toml -o ./proof.np
+cargo instruments --template Allocations --release --bin provekit-cli prove -p ./prover.pkp -i ./Prover.toml -o ./proof.np
 ```
 
 Samply tool [website](https://github.com/mstange/samply/) with instructions to install. It will start local server and
@@ -158,7 +162,7 @@ open a webpage with interactive app to view results. This does not require to ru
 with profiling enabled.
 
 ```sh
-samply record -r 10000 -- ./../../target/release/provekit-cli prove ./prover.pkp ./Prover.toml -o ./proof.np
+samply record -r 10000 -- ./../../target/release/provekit-cli prove -p ./prover.pkp -i ./Prover.toml -o ./proof.np
 ```
 
 ## Benchmarking
