+++ tooling/cli/src/cmd/mod.rs
@@ -1,10 +1,10 @@
 mod analyze_pkp;
 mod circuit_stats;
-mod compile;
 mod generate_gnark_inputs;
 mod prepare;
 mod prove;
 mod show_inputs;
+mod util;
 mod verify;
 
 use {anyhow::Result, argh::FromArgs};
@@ -13,7 +13,7 @@ pub trait Command {
     fn run(&self) -> Result<()>;
 }
 
-/// Prove & verify a compiled Noir program using R1CS.
+/// Compile, prove, and verify Noir programs using R1CS.
 #[derive(FromArgs, PartialEq, Debug)]
 pub struct Args {
     #[argh(subcommand)]
@@ -40,7 +40,6 @@ pub struct Args {
 #[argh(subcommand)]
 enum Commands {
     AnalyzePkp(analyze_pkp::Args),
-    Compile(compile::Args),
     Prepare(prepare::Args),
     Prove(prove::Args),
     CircuitStats(circuit_stats::Args),
@@ -59,7 +58,6 @@ impl Command for Commands {
     fn run(&self) -> Result<()> {
         match self {
             Self::AnalyzePkp(args) => args.run(),
-            Self::Compile(args) => args.run(),
             Self::Prepare(args) => args.run(),
             Self::Prove(args) => args.run(),
             Self::CircuitStats(args) => args.run(),
