+++ tooling/cli/src/cmd/verify.rs
@@ -1,5 +1,5 @@
 use {
-    super::Command,
+    super::{util::resolve_key_path, Command},
     anyhow::{Context, Result},
     argh::FromArgs,
     provekit_common::{file::read, NoirProof, Verifier},
@@ -8,31 +8,32 @@ use {
     tracing::instrument,
 };
 
-/// Verify a Noir proof
+/// Verify a Noir proof.
 #[derive(FromArgs, PartialEq, Eq, Debug)]
 #[argh(subcommand, name = "verify")]
 pub struct Args {
-    /// path to the compiled Noir program
-    #[argh(positional)]
-    verifier_path: PathBuf,
+    /// path to the verifier key (default: <circuit>.pkv)
+    #[argh(option, long = "verifier", short = 'v')]
+    verifier_path: Option<PathBuf>,
 
     /// path to the proof file
-    #[argh(positional)]
+    #[argh(option, long = "proof", default = "PathBuf::from(\"./proof.np\")")]
     proof_path: PathBuf,
 }
 
 impl Command for Args {
     #[instrument(skip_all)]
     fn run(&self) -> Result<()> {
-        // Load verifier and proof in parallel (independent I/O + decompression)
+        let verifier_path = resolve_key_path(self.verifier_path.as_deref(), "pkv")?;
+
+        // Independent I/O + decompression — load both in parallel.
         let (verifier, proof) = rayon::join(
-            || read::<Verifier>(&self.verifier_path).context("while reading Provekit Verifier"),
+            || read::<Verifier>(&verifier_path).context("while reading Provekit Verifier"),
             || read::<NoirProof>(&self.proof_path).context("while reading proof"),
         );
         let mut verifier = verifier?;
         let proof = proof?;
 
-        // Verify the proof
         verifier
             .verify(&proof)
             .context("While verifying Noir proof")?;
